coral-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.3.36
langchain-community>=0.1.0
langchain-mcp-adapters==0.0.3
sseclient-py>=1.7.2
python-dotenv==1.0.0
pydantic>=2.0.0
